first initial

+++ controllers/clothingItems.js
@@ -34,9 +34,15 @@ const deleteItem = (req, res) => {
 
   ClothingItem.findById(itemId)
     .orFail()
-    .then((item) =>
-      item.deleteOne().then(() => res.send({ message: "Item deleted" }))
-    )
+    .then((item) => {
+      if (item.owner.toString() !== req.user._id) {
+        return res
+          .status(403)
+          .send({ message: "You can only delete your own items" });
+      }
+
+      return item.deleteOne().then(() => res.send({ message: "Item deleted" }));
+    })
     .catch((err) => {
       console.error(err);
       if (err.name === "DocumentNotFoundError") {
